requirements.txt
# Only list production dependencies. Everything else can be added to requirements-test.txt and requirements-dev.txt.

# required for cloud2sql to run
SQLAlchemy==1.4.45
PyYAML>=6.0
rich>=12.6.0
resotoclient>=1.2.1
posthog>=2.2.0
requests>=2.28.1

# bundle popular db drivers
psycopg2-binary>=2.9.5 # postgres
pymysql>=1.0.2 # mysql + mariadb

# Install snowflake connector
# Note: this version does not support Python 3.11
# See: https://github.com/snowflakedb/snowflake-connector-python/pull/1349 for reference
snowflake-sqlalchemy>=1.4.5

resotolib>=3.0.0
# all collector plugins
resoto-plugin-aws>=3.0.0
resoto-plugin-digitalocean>=3.0.0
resoto-plugin-gcp>=3.0.0
resoto-plugin-k8s>=3.0.0
